requirements.txt
streamlit>=1.32.0
pyyaml>=6.0
fpdf2>=2.7.0
